refactor auth card
