+++ lib/src/widgets/auth_card.dart
@@ -1,3 +1,5 @@
+library auth_card;
+
 import 'dart:math';
 
 import 'package:another_transformer_page_view/another_transformer_page_view.dart';
@@ -5,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_login/src/models/login_user_type.dart';
 import 'package:provider/provider.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-
+import 'package:flutter_login/src/utils/text_field_utils.dart';
 import 'animated_button.dart';
 import 'animated_icon.dart';
 import 'animated_text.dart';
@@ -13,32 +15,34 @@ import 'animated_text_form_field.dart';
 import 'custom_page_transformer.dart';
 import 'expandable_container.dart';
 import 'fade_in.dart';
-import '../constants.dart';
-import '../providers/auth.dart';
-import '../providers/login_messages.dart';
-import '../providers/login_theme.dart';
-import '../models/login_data.dart';
-import '../dart_helper.dart';
-import '../matrix.dart';
-import '../paddings.dart';
-import '../widget_helper.dart';
+import 'package:flutter_login/src/constants.dart';
+import 'package:flutter_login/src/providers/auth.dart';
+import 'package:flutter_login/src/providers/login_messages.dart';
+import 'package:flutter_login/src/providers/login_theme.dart';
+import 'package:flutter_login/src/models/login_data.dart';
+import 'package:flutter_login/src/dart_helper.dart';
+import 'package:flutter_login/src/matrix.dart';
+import 'package:flutter_login/src/paddings.dart';
+import 'package:flutter_login/src/widget_helper.dart';
 
-// TODO Improvement: Keep just this in auth_card.dart
+part 'login_card.dart';
+part 'recover_card.dart';
+
 class AuthCard extends StatefulWidget {
-  AuthCard(
-      {Key? key,
-      required this.userType,
-      this.padding = const EdgeInsets.all(0),
-      this.loadingController,
-      this.userValidator,
-      this.passwordValidator,
-      this.onSubmit,
-      this.onSubmitCompleted,
-      this.hideForgotPasswordButton = false,
-      this.hideSignUpButton = false,
-      this.loginAfterSignUp = true,
-      this.hideProvidersTitle = false})
-      : super(key: key);
+  AuthCard({
+    Key? key,
+    required this.userType,
+    this.padding = const EdgeInsets.all(0),
+    this.loadingController,
+    this.userValidator,
+    this.passwordValidator,
+    this.onSubmit,
+    this.onSubmitCompleted,
+    this.hideForgotPasswordButton = false,
+    this.hideSignUpButton = false,
+    this.loginAfterSignUp = true,
+    this.hideProvidersTitle = false,
+  }) : super(key: key);
 
   final EdgeInsets padding;
   final AnimationController? loadingController;
@@ -342,717 +346,6 @@ class AuthCardState extends State<AuthCard> with TickerProviderStateMixin {
           child: current,
         );
       },
-    );
-  }
-}
-
-// TODO Improvement: Modularize this in a login_card.dart
-class _LoginCard extends StatefulWidget {
-  _LoginCard(
-      {Key? key,
-      this.loadingController,
-      required this.userValidator,
-      required this.passwordValidator,
-      required this.onSwitchRecoveryPassword,
-      required this.userType,
-      this.onSwitchAuth,
-      this.onSubmitCompleted,
-      this.hideForgotPasswordButton = false,
-      this.hideSignUpButton = false,
-      this.loginAfterSignUp = true,
-      this.hideProvidersTitle = false})
-      : super(key: key);
-
-  final AnimationController? loadingController;
-  final FormFieldValidator<String>? userValidator;
-  final FormFieldValidator<String>? passwordValidator;
-  final Function onSwitchRecoveryPassword;
-  final Function? onSwitchAuth;
-  final Function? onSubmitCompleted;
-  final bool hideForgotPasswordButton;
-  final bool hideSignUpButton;
-  final bool loginAfterSignUp;
-  final bool hideProvidersTitle;
-  final LoginUserType userType;
-
-  @override
-  _LoginCardState createState() => _LoginCardState();
-}
-
-class _LoginCardState extends State<_LoginCard> with TickerProviderStateMixin {
-  final GlobalKey<FormState> _formKey = GlobalKey();
-
-  final _passwordFocusNode = FocusNode();
-  final _confirmPasswordFocusNode = FocusNode();
-
-  TextEditingController? _nameController;
-  TextEditingController? _passController;
-  TextEditingController? _confirmPassController;
-
-  var _isLoading = false;
-  var _isSubmitting = false;
-  var _showShadow = true;
-
-  /// switch between login and signup
-  late AnimationController _loadingController;
-  late AnimationController _switchAuthController;
-  late AnimationController _postSwitchAuthController;
-  late AnimationController _submitController;
-
-  ///list of AnimationController each one responsible for a authentication provider icon
-  List<AnimationController> _providerControllerList = <AnimationController>[];
-
-  Interval? _nameTextFieldLoadingAnimationInterval;
-  Interval? _passTextFieldLoadingAnimationInterval;
-  Interval? _textButtonLoadingAnimationInterval;
-  late Animation<double> _buttonScaleAnimation;
-
-  bool get buttonEnabled => !_isLoading && !_isSubmitting;
-
-  @override
-  void initState() {
-    super.initState();
-
-    final auth = Provider.of<Auth>(context, listen: false);
-    _nameController = TextEditingController(text: auth.email);
-    _passController = TextEditingController(text: auth.password);
-    _confirmPassController = TextEditingController(text: auth.confirmPassword);
-
-    _loadingController = widget.loadingController ??
-        (AnimationController(
-          vsync: this,
-          duration: Duration(milliseconds: 1150),
-          reverseDuration: Duration(milliseconds: 300),
-        )..value = 1.0);
-
-    _loadingController.addStatusListener(handleLoadingAnimationStatus);
-
-    _switchAuthController = AnimationController(
-      vsync: this,
-      duration: Duration(milliseconds: 800),
-    );
-    _postSwitchAuthController = AnimationController(
-      vsync: this,
-      duration: Duration(milliseconds: 150),
-    );
-    _submitController = AnimationController(
-      vsync: this,
-      duration: Duration(milliseconds: 1000),
-    );
-    _providerControllerList = auth.loginProviders
-        .map(
-          (e) => AnimationController(
-            vsync: this,
-            duration: Duration(milliseconds: 1000),
-          ),
-        )
-        .toList();
-
-    _nameTextFieldLoadingAnimationInterval = const Interval(0, .85);
-    _passTextFieldLoadingAnimationInterval = const Interval(.15, 1.0);
-    _textButtonLoadingAnimationInterval =
-        const Interval(.6, 1.0, curve: Curves.easeOut);
-    _buttonScaleAnimation =
-        Tween<double>(begin: 0.0, end: 1.0).animate(CurvedAnimation(
-      parent: _loadingController,
-      curve: Interval(.4, 1.0, curve: Curves.easeOutBack),
-    ));
-  }
-
-  void handleLoadingAnimationStatus(AnimationStatus status) {
-    if (status == AnimationStatus.forward) {
-      setState(() => _isLoading = true);
-    }
-    if (status == AnimationStatus.completed) {
-      setState(() => _isLoading = false);
-    }
-  }
-
-  @override
-  void dispose() {
-    _loadingController.removeStatusListener(handleLoadingAnimationStatus);
-    _passwordFocusNode.dispose();
-    _confirmPasswordFocusNode.dispose();
-
-    _switchAuthController.dispose();
-    _postSwitchAuthController.dispose();
-    _submitController.dispose();
-
-    _providerControllerList.forEach((controller) {
-      controller.dispose();
-    });
-    super.dispose();
-  }
-
-  void _switchAuthMode() {
-    final auth = Provider.of<Auth>(context, listen: false);
-    final newAuthMode = auth.switchAuth();
-
-    if (newAuthMode == AuthMode.Signup) {
-      _switchAuthController.forward();
-    } else {
-      _switchAuthController.reverse();
-    }
-  }
-
-  Future<bool> _submit() async {
-    // a hack to force unfocus the soft keyboard. If not, after change-route
-    // animation completes, it will trigger rebuilding this widget and show all
-    // textfields and buttons again before going to new route
-    FocusScope.of(context).requestFocus(FocusNode());
-
-    final messages = Provider.of<LoginMessages>(context, listen: false);
-
-    if (!_formKey.currentState!.validate()) {
-      return false;
-    }
-
-    _formKey.currentState!.save();
-    await _submitController.forward();
-    setState(() => _isSubmitting = true);
-    final auth = Provider.of<Auth>(context, listen: false);
-    String? error;
-
-    if (auth.isLogin) {
-      error = await auth.onLogin!(LoginData(
-        name: auth.email,
-        password: auth.password,
-      ));
-    } else {
-      error = await auth.onSignup!(LoginData(
-        name: auth.email,
-        password: auth.password,
-      ));
-    }
-
-    // workaround to run after _cardSizeAnimation in parent finished
-    // need a cleaner way but currently it works so..
-    Future.delayed(const Duration(milliseconds: 270), () {
-      setState(() => _showShadow = false);
-    });
-
-    await _submitController.reverse();
-
-    if (!DartHelper.isNullOrEmpty(error)) {
-      showErrorToast(context, messages.flushbarTitleError, error!);
-      Future.delayed(const Duration(milliseconds: 271), () {
-        setState(() => _showShadow = true);
-      });
-      setState(() => _isSubmitting = false);
-      return false;
-    }
-
-    if (auth.isSignup && !widget.loginAfterSignUp) {
-      showSuccessToast(
-          context, messages.flushbarTitleSuccess, messages.signUpSuccess);
-      _switchAuthMode();
-      setState(() => _isSubmitting = false);
-      return false;
-    }
-
-    widget.onSubmitCompleted!();
-
-    return true;
-  }
-
-  Future<bool> _loginProviderSubmit(
-      {required AnimationController control,
-      required ProviderAuthCallback callback}) async {
-    await control.forward();
-
-    String? error;
-
-    error = await callback();
-
-    // workaround to run after _cardSizeAnimation in parent finished
-    // need a cleaner way but currently it works so..
-    Future.delayed(const Duration(milliseconds: 270), () {
-      setState(() => _showShadow = false);
-    });
-
-    await control.reverse();
-
-    final messages = Provider.of<LoginMessages>(context, listen: false);
-
-    if (!DartHelper.isNullOrEmpty(error)) {
-      showErrorToast(context, messages.flushbarTitleError, error!);
-      Future.delayed(const Duration(milliseconds: 271), () {
-        setState(() => _showShadow = true);
-      });
-      return false;
-    }
-
-    widget.onSubmitCompleted!();
-
-    return true;
-  }
-
-  // TODO Improvement: Common function to login_card.dart and recover_card.dart
-  // Create a resource to import these function and avoid duplicated code
-  String _getAutofillHints(LoginUserType userType) {
-    switch (userType) {
-      case LoginUserType.name:
-        return AutofillHints.username;
-      case LoginUserType.phone:
-        return AutofillHints.telephoneNumber;
-      case LoginUserType.email:
-      default:
-        return AutofillHints.email;
-    }
-  }
-
-  // TODO Improvement: Common function to login_card.dart and recover_card.dart
-  // Create a resource to import these function and avoid duplicated code
-  TextInputType _getKeyboardType(LoginUserType userType) {
-    switch (userType) {
-      case LoginUserType.name:
-        return TextInputType.name;
-      case LoginUserType.phone:
-        return TextInputType.number;
-      case LoginUserType.email:
-      default:
-        return TextInputType.emailAddress;
-    }
-  }
-
-  Widget _buildUserField(
-    double width,
-    LoginMessages messages,
-    Auth auth,
-  ) {
-    return AnimatedTextFormField(
-      controller: _nameController,
-      width: width,
-      loadingController: _loadingController,
-      interval: _nameTextFieldLoadingAnimationInterval,
-      labelText: messages.userHint,
-      autofillHints: [_getAutofillHints(widget.userType)],
-      prefixIcon: Icon(FontAwesomeIcons.solidUserCircle),
-      keyboardType: _getKeyboardType(widget.userType),
-      textInputAction: TextInputAction.next,
-      onFieldSubmitted: (value) {
-        FocusScope.of(context).requestFocus(_passwordFocusNode);
-      },
-      validator: widget.userValidator,
-      onSaved: (value) => auth.email = value!,
-    );
-  }
-
-  Widget _buildPasswordField(double width, LoginMessages messages, Auth auth) {
-    return AnimatedPasswordTextFormField(
-      animatedWidth: width,
-      loadingController: _loadingController,
-      interval: _passTextFieldLoadingAnimationInterval,
-      labelText: messages.passwordHint,
-      autofillHints:
-          auth.isLogin ? [AutofillHints.password] : [AutofillHints.newPassword],
-      controller: _passController,
-      textInputAction:
-          auth.isLogin ? TextInputAction.done : TextInputAction.next,
-      focusNode: _passwordFocusNode,
-      onFieldSubmitted: (value) {
-        if (auth.isLogin) {
-          _submit();
-        } else {
-          // SignUp
-          FocusScope.of(context).requestFocus(_confirmPasswordFocusNode);
-        }
-      },
-      validator: widget.passwordValidator,
-      onSaved: (value) => auth.password = value!,
-    );
-  }
-
-  Widget _buildConfirmPasswordField(
-      double width, LoginMessages messages, Auth auth) {
-    return AnimatedPasswordTextFormField(
-      animatedWidth: width,
-      enabled: auth.isSignup,
-      loadingController: _loadingController,
-      inertiaController: _postSwitchAuthController,
-      inertiaDirection: TextFieldInertiaDirection.right,
-      labelText: messages.confirmPasswordHint,
-      controller: _confirmPassController,
-      textInputAction: TextInputAction.done,
-      focusNode: _confirmPasswordFocusNode,
-      onFieldSubmitted: (value) => _submit(),
-      validator: auth.isSignup
-          ? (value) {
-              if (value != _passController!.text) {
-                return messages.confirmPasswordError;
-              }
-              return null;
-            }
-          : (value) => null,
-      onSaved: (value) => auth.confirmPassword = value!,
-    );
-  }
-
-  Widget _buildForgotPassword(ThemeData theme, LoginMessages messages) {
-    return FadeIn(
-      controller: _loadingController,
-      fadeDirection: FadeDirection.bottomToTop,
-      offset: .5,
-      curve: _textButtonLoadingAnimationInterval,
-      child: TextButton(
-        onPressed: buttonEnabled
-            ? () {
-                // save state to populate email field on recovery card
-                _formKey.currentState!.save();
-                widget.onSwitchRecoveryPassword();
-              }
-            : null,
-        child: Text(
-          messages.forgotPasswordButton,
-          style: theme.textTheme.bodyText2,
-          textAlign: TextAlign.left,
-        ),
-      ),
-    );
-  }
-
-  Widget _buildSubmitButton(
-      ThemeData theme, LoginMessages messages, Auth auth) {
-    return ScaleTransition(
-      scale: _buttonScaleAnimation,
-      child: AnimatedButton(
-        controller: _submitController,
-        text: auth.isLogin ? messages.loginButton : messages.signupButton,
-        onPressed: _submit,
-      ),
-    );
-  }
-
-  Widget _buildSwitchAuthButton(ThemeData theme, LoginMessages messages,
-      Auth auth, LoginTheme loginTheme) {
-    return FadeIn(
-      controller: _loadingController,
-      offset: .5,
-      curve: _textButtonLoadingAnimationInterval,
-      fadeDirection: FadeDirection.topToBottom,
-      child: MaterialButton(
-        disabledTextColor: theme.primaryColor,
-        onPressed: buttonEnabled ? _switchAuthMode : null,
-        padding: loginTheme.authButtonPadding ??
-            EdgeInsets.symmetric(horizontal: 30.0, vertical: 8.0),
-        materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-        textColor: theme.primaryColor,
-        child: AnimatedText(
-          text: auth.isSignup ? messages.loginButton : messages.signupButton,
-          textRotation: AnimatedTextRotation.down,
-        ),
-      ),
-    );
-  }
-
-  Widget _buildProvidersLogInButton(ThemeData theme, LoginMessages messages,
-      Auth auth, LoginTheme loginTheme) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: auth.loginProviders.map((loginProvider) {
-        var index = auth.loginProviders.indexOf(loginProvider);
-        return Padding(
-          padding: loginTheme.providerButtonPadding ??
-              const EdgeInsets.symmetric(horizontal: 6.0, vertical: 8.0),
-          child: ScaleTransition(
-            scale: _buttonScaleAnimation,
-            child: AnimatedIconButton(
-              icon: loginProvider.icon,
-              controller: _providerControllerList[index],
-              tooltip: '',
-              onPressed: () => _loginProviderSubmit(
-                control: _providerControllerList[index],
-                callback: () {
-                  return loginProvider.callback();
-                },
-              ),
-            ),
-          ),
-        );
-      }).toList(),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final auth = Provider.of<Auth>(context, listen: true);
-    final isLogin = auth.isLogin;
-    final messages = Provider.of<LoginMessages>(context, listen: false);
-    final loginTheme = Provider.of<LoginTheme>(context, listen: false);
-    final theme = Theme.of(context);
-    final deviceSize = MediaQuery.of(context).size;
-    final cardWidth = min(deviceSize.width * 0.75, 360.0);
-    const cardPadding = 16.0;
-    final textFieldWidth = cardWidth - cardPadding * 2;
-    final authForm = Form(
-      key: _formKey,
-      child: Column(
-        children: [
-          Container(
-            padding: EdgeInsets.only(
-              left: cardPadding,
-              right: cardPadding,
-              top: cardPadding + 10,
-            ),
-            width: cardWidth,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: <Widget>[
-                _buildUserField(textFieldWidth, messages, auth),
-                SizedBox(height: 20),
-                _buildPasswordField(textFieldWidth, messages, auth),
-                SizedBox(height: 10),
-              ],
-            ),
-          ),
-          ExpandableContainer(
-            backgroundColor: theme.accentColor,
-            controller: _switchAuthController,
-            initialState: isLogin
-                ? ExpandableContainerState.shrunk
-                : ExpandableContainerState.expanded,
-            alignment: Alignment.topLeft,
-            color: theme.cardTheme.color,
-            width: cardWidth,
-            padding: EdgeInsets.symmetric(
-              horizontal: cardPadding,
-              vertical: 10,
-            ),
-            onExpandCompleted: () => _postSwitchAuthController.forward(),
-            child: _buildConfirmPasswordField(textFieldWidth, messages, auth),
-          ),
-          Container(
-            padding: Paddings.fromRBL(cardPadding),
-            width: cardWidth,
-            child: Column(
-              children: <Widget>[
-                !widget.hideForgotPasswordButton
-                    ? _buildForgotPassword(theme, messages)
-                    : SizedBox.fromSize(
-                        size: Size.fromHeight(16),
-                      ),
-                _buildSubmitButton(theme, messages, auth),
-                !widget.hideSignUpButton
-                    ? _buildSwitchAuthButton(theme, messages, auth, loginTheme)
-                    : SizedBox.fromSize(
-                        size: Size.fromHeight(10),
-                      ),
-                // TODO: Fix animation size when ending. See #184
-                // auth.loginProviders.isNotEmpty && !widget.hideProvidersTitle
-                //     ? Row(children: <Widget>[
-                //         Expanded(child: Divider()),
-                //         Padding(
-                //           padding: const EdgeInsets.all(8.0),
-                //           child: Text(messages.providersTitle),
-                //         ),
-                //         Expanded(child: Divider()),
-                //       ])
-                //     : Container(),
-                _buildProvidersLogInButton(theme, messages, auth, loginTheme),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-
-    return FittedBox(
-      child: Card(
-        elevation: _showShadow ? theme.cardTheme.elevation : 0,
-        child: authForm,
-      ),
-    );
-  }
-}
-
-// TODO Improvement: Modularize this in a recover_card.dart
-class _RecoverCard extends StatefulWidget {
-  _RecoverCard({
-    Key? key,
-    required this.userValidator,
-    required this.onSwitchLogin,
-    required this.userType,
-  }) : super(key: key);
-
-  final FormFieldValidator<String>? userValidator;
-  final Function onSwitchLogin;
-  final LoginUserType userType;
-
-  @override
-  _RecoverCardState createState() => _RecoverCardState();
-}
-
-class _RecoverCardState extends State<_RecoverCard>
-    with SingleTickerProviderStateMixin {
-  final GlobalKey<FormState> _formRecoverKey = GlobalKey();
-
-  TextEditingController? _nameController;
-
-  var _isSubmitting = false;
-
-  AnimationController? _submitController;
-
-  @override
-  void initState() {
-    super.initState();
-
-    final auth = Provider.of<Auth>(context, listen: false);
-    _nameController = TextEditingController(text: auth.email);
-
-    _submitController = AnimationController(
-      vsync: this,
-      duration: Duration(milliseconds: 1000),
-    );
-  }
-
-  @override
-  void dispose() {
-    _submitController!.dispose();
-    super.dispose();
-  }
-
-  Future<bool> _submit() async {
-    if (!_formRecoverKey.currentState!.validate()) {
-      return false;
-    }
-    final auth = Provider.of<Auth>(context, listen: false);
-    final messages = Provider.of<LoginMessages>(context, listen: false);
-
-    _formRecoverKey.currentState!.save();
-    await _submitController!.forward();
-    setState(() => _isSubmitting = true);
-    final error = await auth.onRecoverPassword!(auth.email);
-
-    if (error != null) {
-      showErrorToast(context, messages.flushbarTitleError, error);
-      setState(() => _isSubmitting = false);
-      await _submitController!.reverse();
-      return false;
-    } else {
-      showSuccessToast(context, messages.flushbarTitleSuccess,
-          messages.recoverPasswordSuccess);
-      setState(() => _isSubmitting = false);
-      await _submitController!.reverse();
-      return true;
-    }
-  }
-
-  // TODO Improvement: Common function to login_card.dart and recover_card.dart
-  // Create a resource to import these function and avoid duplicated code
-  String _getAutofillHints(LoginUserType userType) {
-    switch (userType) {
-      case LoginUserType.name:
-        return AutofillHints.username;
-      case LoginUserType.phone:
-        return AutofillHints.telephoneNumber;
-      case LoginUserType.email:
-      default:
-        return AutofillHints.email;
-    }
-  }
-
-  // TODO Improvement: Common function to login_card.dart and recover_card.dart
-  // Create a resource to import these function and avoid duplicated code
-  TextInputType _getKeyboardType(LoginUserType userType) {
-    switch (userType) {
-      case LoginUserType.name:
-        return TextInputType.name;
-      case LoginUserType.phone:
-        return TextInputType.number;
-      case LoginUserType.email:
-      default:
-        return TextInputType.emailAddress;
-    }
-  }
-
-  Widget _buildRecoverNameField(
-      double width, LoginMessages messages, Auth auth) {
-    return AnimatedTextFormField(
-      controller: _nameController,
-      width: width,
-      labelText: messages.userHint,
-      prefixIcon: Icon(FontAwesomeIcons.solidUserCircle),
-      keyboardType: _getKeyboardType(widget.userType),
-      autofillHints: [_getAutofillHints(widget.userType)],
-      textInputAction: TextInputAction.done,
-      onFieldSubmitted: (value) => _submit(),
-      validator: widget.userValidator,
-      onSaved: (value) => auth.email = value!,
-    );
-  }
-
-  Widget _buildRecoverButton(ThemeData theme, LoginMessages messages) {
-    return AnimatedButton(
-      controller: _submitController,
-      text: messages.recoverPasswordButton,
-      onPressed: !_isSubmitting ? _submit : null,
-    );
-  }
-
-  Widget _buildBackButton(ThemeData theme, LoginMessages messages) {
-    return MaterialButton(
-      onPressed: !_isSubmitting
-          ? () {
-              _formRecoverKey.currentState!.save();
-              widget.onSwitchLogin();
-            }
-          : null,
-      padding: EdgeInsets.symmetric(horizontal: 30.0, vertical: 4),
-      materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-      textColor: theme.primaryColor,
-      child: Text(messages.goBackButton),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final auth = Provider.of<Auth>(context, listen: false);
-    final messages = Provider.of<LoginMessages>(context, listen: false);
-    final deviceSize = MediaQuery.of(context).size;
-    final cardWidth = min(deviceSize.width * 0.75, 360.0);
-    const cardPadding = 16.0;
-    final textFieldWidth = cardWidth - cardPadding * 2;
-
-    return FittedBox(
-      // width: cardWidth,
-      child: Card(
-        child: Container(
-          padding: const EdgeInsets.only(
-            left: cardPadding,
-            top: cardPadding + 10.0,
-            right: cardPadding,
-            bottom: cardPadding,
-          ),
-          width: cardWidth,
-          alignment: Alignment.center,
-          child: Form(
-            key: _formRecoverKey,
-            child: Column(
-              children: [
-                Text(
-                  messages.recoverPasswordIntro,
-                  key: kRecoverPasswordIntroKey,
-                  textAlign: TextAlign.center,
-                  style: theme.textTheme.bodyText2,
-                ),
-                SizedBox(height: 20),
-                _buildRecoverNameField(textFieldWidth, messages, auth),
-                SizedBox(height: 20),
-                Text(
-                  messages.recoverPasswordDescription,
-                  key: kRecoverPasswordDescriptionKey,
-                  textAlign: TextAlign.center,
-                  style: theme.textTheme.bodyText2,
-                ),
-                SizedBox(height: 26),
-                _buildRecoverButton(theme, messages),
-                _buildBackButton(theme, messages),
-              ],
-            ),
-          ),
-        ),
-      ),
     );
   }
 }
