+++ lib/src/widgets/recover_card.dart
@@ -1,0 +1,165 @@
+part of auth_card;
+
+// TODO Improvement: Modularize this in a recover_card.dart
+class _RecoverCard extends StatefulWidget {
+  _RecoverCard({
+    Key? key,
+    required this.userValidator,
+    required this.onSwitchLogin,
+    required this.userType,
+  }) : super(key: key);
+
+  final FormFieldValidator<String>? userValidator;
+  final Function onSwitchLogin;
+  final LoginUserType userType;
+
+  @override
+  _RecoverCardState createState() => _RecoverCardState();
+}
+
+class _RecoverCardState extends State<_RecoverCard>
+    with SingleTickerProviderStateMixin {
+  final GlobalKey<FormState> _formRecoverKey = GlobalKey();
+
+  TextEditingController? _nameController;
+
+  var _isSubmitting = false;
+
+  AnimationController? _submitController;
+
+  @override
+  void initState() {
+    super.initState();
+
+    final auth = Provider.of<Auth>(context, listen: false);
+    _nameController = TextEditingController(text: auth.email);
+
+    _submitController = AnimationController(
+      vsync: this,
+      duration: Duration(milliseconds: 1000),
+    );
+  }
+
+  @override
+  void dispose() {
+    _submitController!.dispose();
+    super.dispose();
+  }
+
+  Future<bool> _submit() async {
+    if (!_formRecoverKey.currentState!.validate()) {
+      return false;
+    }
+    final auth = Provider.of<Auth>(context, listen: false);
+    final messages = Provider.of<LoginMessages>(context, listen: false);
+
+    _formRecoverKey.currentState!.save();
+    await _submitController!.forward();
+    setState(() => _isSubmitting = true);
+    final error = await auth.onRecoverPassword!(auth.email);
+
+    if (error != null) {
+      showErrorToast(context, messages.flushbarTitleError, error);
+      setState(() => _isSubmitting = false);
+      await _submitController!.reverse();
+      return false;
+    } else {
+      showSuccessToast(context, messages.flushbarTitleSuccess,
+          messages.recoverPasswordSuccess);
+      setState(() => _isSubmitting = false);
+      await _submitController!.reverse();
+      return true;
+    }
+  }
+
+  Widget _buildRecoverNameField(
+      double width, LoginMessages messages, Auth auth) {
+    return AnimatedTextFormField(
+      controller: _nameController,
+      width: width,
+      labelText: messages.userHint,
+      prefixIcon: Icon(FontAwesomeIcons.solidUserCircle),
+      keyboardType: TextFieldUtils.getKeyboardType(widget.userType),
+      autofillHints: [TextFieldUtils.getAutofillHints(widget.userType)],
+      textInputAction: TextInputAction.done,
+      onFieldSubmitted: (value) => _submit(),
+      validator: widget.userValidator,
+      onSaved: (value) => auth.email = value!,
+    );
+  }
+
+  Widget _buildRecoverButton(ThemeData theme, LoginMessages messages) {
+    return AnimatedButton(
+      controller: _submitController,
+      text: messages.recoverPasswordButton,
+      onPressed: !_isSubmitting ? _submit : null,
+    );
+  }
+
+  Widget _buildBackButton(ThemeData theme, LoginMessages messages) {
+    return MaterialButton(
+      onPressed: !_isSubmitting
+          ? () {
+              _formRecoverKey.currentState!.save();
+              widget.onSwitchLogin();
+            }
+          : null,
+      padding: EdgeInsets.symmetric(horizontal: 30.0, vertical: 4),
+      materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
+      textColor: theme.primaryColor,
+      child: Text(messages.goBackButton),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final auth = Provider.of<Auth>(context, listen: false);
+    final messages = Provider.of<LoginMessages>(context, listen: false);
+    final deviceSize = MediaQuery.of(context).size;
+    final cardWidth = min(deviceSize.width * 0.75, 360.0);
+    const cardPadding = 16.0;
+    final textFieldWidth = cardWidth - cardPadding * 2;
+
+    return FittedBox(
+      // width: cardWidth,
+      child: Card(
+        child: Container(
+          padding: const EdgeInsets.only(
+            left: cardPadding,
+            top: cardPadding + 10.0,
+            right: cardPadding,
+            bottom: cardPadding,
+          ),
+          width: cardWidth,
+          alignment: Alignment.center,
+          child: Form(
+            key: _formRecoverKey,
+            child: Column(
+              children: [
+                Text(
+                  messages.recoverPasswordIntro,
+                  key: kRecoverPasswordIntroKey,
+                  textAlign: TextAlign.center,
+                  style: theme.textTheme.bodyText2,
+                ),
+                SizedBox(height: 20),
+                _buildRecoverNameField(textFieldWidth, messages, auth),
+                SizedBox(height: 20),
+                Text(
+                  messages.recoverPasswordDescription,
+                  key: kRecoverPasswordDescriptionKey,
+                  textAlign: TextAlign.center,
+                  style: theme.textTheme.bodyText2,
+                ),
+                SizedBox(height: 26),
+                _buildRecoverButton(theme, messages),
+                _buildBackButton(theme, messages),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
